divide main.js by modules
